refactor: [Issuer Adapter] move common test code to a new file
